backend/requirements/development.txt
-r base.txt
black==25.11.0
flake8==7.3.0
ipython==8.18.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
